requirements.txt
openai==2.28.0
requests==2.32.5
python-dotenv==1.2.2